requirements.txt
# Streamlined Whisper Transcriber Requirements
# Mobile-first, modern transcription service
# ================================================

# --- Core Web Framework ---
fastapi>=0.111.0
uvicorn[standard]>=0.29.0

# --- Database (SQLite for simplicity) ---
sqlalchemy>=2.0.30
alembic>=1.12.0

# --- Security & Configuration ---
cryptography>=41.0.0
python-magic>=0.4.27
filetype>=1.2.0
bcrypt>=4.0.1

# --- File Handling ---
python-multipart>=0.0.6
aiofiles>=23.2.1

# --- Audio Processing & Transcription ---
openai-whisper==20240930
torch>=2.2.2
pydub>=0.25.1
# Audio enhancement dependencies
librosa>=0.10.0
soundfile>=0.12.1
numpy>=1.24.0
scipy>=1.10.0

# --- Background Job Processing ---
celery>=5.3
redis>=5.0.0

# --- File Type Validation ---
python-magic>=0.4.27

# --- Authentication (Simplified) ---
PyJWT>=2.8.0
passlib[bcrypt]>=1.7.4

# --- Configuration ---
python-dotenv>=1.1.0
pydantic[dotenv]>=2.6.0
tomli>=2.0.0  # TOML parsing for Python < 3.11

# --- System Monitoring ---
psutil>=5.9.0

# --- Backup & Recovery System ---
# High-performance compression
zstandard>=0.22.0
# File system monitoring for real-time backup
watchdog>=4.0.0
# Job scheduling for automated backups
schedule>=1.2.0
# Optional: S3-compatible storage (AWS, MinIO, etc.)
boto3>=1.34.0
# Optional: SFTP/SSH remote storage
paramiko>=3.4.0

# ================================================
# Total: 18 core packages (6 added for backup system)
# Backup system features:
# - Real-time file monitoring and backup
# - SQLite point-in-time recovery
# - ZSTD compression with gzip fallback  
# - Multiple storage backends (local, S3, SFTP)
# - Automated scheduling and retention
# - Disaster recovery procedures
# ================================================

# System requirements (install separately):
# sudo apt-get install -y ffmpeg libmagic1